requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-DebugToolbar==0.9.0
Flask-Login==0.2.11
Flask-OpenID==1.2.1
Flask-SQLAlchemy==1.0
Flask-Uploads==0.1.3
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.5.0
SQLAlchemy==0.9.6
Werkzeug==0.9.6
alembic==0.6.5
argparse==1.2.1
blinker==1.3
itsdangerous==0.24
jsmin==2.0.9
python-openid==2.2.5
steamodd==4.10
webassets==0.10
wsgiref==0.1.2